Qt'ifize the editor
Work in progress

I have created a window with an almost empty menu which displays
the map in a QGLWidget.

Most of the input doesn't work, I've only ported "Show and remove grid"
with 'g' and 'h' to show how it should be done.

All the Qt stuff from main.cpp should be moved to a QApplication subclass